refactor(gateway): extract response_types.go from server.go
Phase 3b file split: moves ProcessingResponse, StatusCreated/StatusDeduplicated,
BatchProcessingResponse, ProcessingResult, BatchSummary, and the
NewDuplicateResponseFromRR/NewCRDCreatedResponse factory functions into their
own file. Pure structural move: 154 Ginkgo specs still pass.
